incorporated redux with next-redux-wrapper
