+++ frontend/package.json
@@ -15,8 +15,11 @@
     "next": "^5.1.0",
     "react": "^16.3.0",
     "react-redux": "^5.0.7",
+    "next-redux-wrapper": "^1.3.5",
+    "react-dom": "^16.0.0",
     "reactstrap": "^6.1.0",
     "redux": "^4.0.0",
+    "redux-thunk": "^2.2.0",
     "redux-persist": "^5.10.0"
   },
   "devDependencies": {
